# Remove unnecessary graphql fields