--- a/graphql/documents/data/tag-slim.graphql
+++ b/graphql/documents/data/tag-slim.graphql
@@ -1,0 +1,5 @@
+fragment SlimTagData on Tag {
+  id
+  name
+  image_path
+}