--- a/graphql/documents/queries/gallery.graphql
+++ b/graphql/documents/queries/gallery.graphql
@@ -2,7 +2,7 @@ query FindGalleries($filter: FindFilterType, $gallery_filter: GalleryFilterType)
   findGalleries(gallery_filter: $gallery_filter, filter: $filter) {
     count
     galleries {
-      ...GallerySlimData
+      ...SlimGalleryData
     }
   }
 }