--- a/ui/v2.5/src/components/Galleries/GalleryDetails/GalleryScenesPanel.tsx
+++ b/ui/v2.5/src/components/Galleries/GalleryDetails/GalleryScenesPanel.tsx
@@ -3,7 +3,7 @@ import * as GQL from "src/core/generated-graphql";
 import { SceneCard } from "src/components/Scenes/SceneCard";
 
 interface IGalleryScenesPanelProps {
-  scenes: GQL.SceneDataFragment[];
+  scenes: GQL.SlimSceneDataFragment[];
 }
 
 export const GalleryScenesPanel: React.FC<IGalleryScenesPanelProps> = ({