Fixes #1368 

Removes fields from graphql queries that are unnecessary. 

Should result in improvements for queries in the following pages:
* galleries
* gallery detail
* image detail
* movie detail
* performers
* performer detail
* scene detail

Also renames `GallerySlimData` to `SlimGalleryData` for consistency.