@WithoutPants 2021-05-07
I have optimised a small number of joins to improve performer and tag resolver queries.

With all of these changes, with my contrived database, I got the following page improvements:

galleries - 12.36s -> 81ms
performers - 1.83s -> 641ms
tags - 1.7s -> 156ms